renamed file to plugin.ts
