+++ src/__tests__/plugin.spec.ts
